修改关键词搜索返回实体类

+++ api/src/main/java/com/toutiao/appV2/model/plot/PlotListRequest.java
@@ -31,7 +31,7 @@ public class PlotListRequest {
 
     @JsonProperty("distance")
     @ApiParam("附近距离")
-    private Integer distance;
+    private double distance;
 
     @JsonProperty("districtId")
     @ApiParam("区域")
